# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Compass Car API",
    "version": "1.0.0",
    "description": "API para um sistema de locação de carros."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/clientes": {
      "post": {
        "summary": "Cria um novo cliente",
        "tags": [
          "Clientes"
        ],
        "description": "Endpoint protegido para cadastrar um novo cliente no sistema. É necessário um token JWT para acessar essa rota.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nomeCompleto",
                  "dataDeNascimento",
                  "cpf",
                  "email",
                  "telefone"
                ],
                "properties": {
                  "nomeCompleto": {
                    "type": "string",
                    "example": "Kelly Slater"
                  },
                  "dataDeNascimento": {
                    "type": "string",
                    "format": "date",
                    "example": "1997-11-11"
                  },
                  "cpf": {
                    "type": "string",
                    "example": "321.654.987-22"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "example.test95@yahoo.com"
                  },
                  "telefone": {
                    "type": "string",
                    "example": "11987654321"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "e7d66f9d-3c10-4e9e-b939-8a3a19159b6d"
                    },
                    "nomeCompleto": {
                      "type": "string",
                      "example": "Kelly Slater"
                    },
                    "dataDeNascimento": {
                      "type": "string",
                      "format": "date",
                      "example": "1997-11-11"
                    },
                    "cpf": {
                      "type": "string",
                      "example": "321.654.987-22"
                    },
                    "email": {
                      "type": "string",
                      "format": "email",
                      "example": "example.test95@yahoo.com"
                    },
                    "telefone": {
                      "type": "string",
                      "example": "11987654321"
                    },
                    "dataCadastro": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2024-10-12T19:34:42.000Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida (dados incorretos ou ausentes)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Dados inválidos: Data de Nascimento inválida"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falha na autenticação JWT (Token ausente ou inválido)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Token inválido ou ausente"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor"
          }
        }
      },
      "get": {
        "summary": "Lista todos os clientes cadastrados (incluindo excluídos ou não)",
        "tags": [
          "Clientes"
        ],
        "description": "Endpoint para listar todos os clientes cadastrados, com a possibilidade de filtrar por nome, email, CPF, telefone e status de exclusão (excluídos ou não). Esta rota é protegida por autenticação via token JWT.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "nomeCompleto",
            "schema": {
              "type": "string"
            },
            "description": "Parte do nome do cliente"
          },
          {
            "in": "query",
            "name": "cpf",
            "schema": {
              "type": "string"
            },
            "description": "CPF completo e válido do cliente"
          },
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string",
              "format": "email"
            },
            "description": "Parte do email do cliente"
          },
          {
            "in": "query",
            "name": "telefone",
            "schema": {
              "type": "string"
            },
            "description": "Parte do telefone do cliente"
          },
          {
            "in": "query",
            "name": "excluido",
            "schema": {
              "type": "string",
              "enum": [
                "sim",
                "não"
              ]
            },
            "description": "Filtrar clientes excluídos (sim) ou não excluídos (não). Valor padrão: \"não\".\n"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 5,
              "maximum": 10
            },
            "description": "Quantidade máxima de clientes por página. (default: 5, min: 5, max: 10)\n"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Página atual da paginação. Valor padrão: 1.\n"
          },
          {
            "in": "query",
            "name": "orderBy",
            "schema": {
              "type": "string",
              "enum": [
                "nomeCompleto",
                "dataCadastro",
                "dataExclusao"
              ]
            },
            "description": "Campo pelo qual a listagem deve ser ordenada. Exemplos: nomeCompleto, dataCadastro, dataExclusao.\n"
          },
          {
            "in": "query",
            "name": "orderDirection",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "Direção da ordenação (ascendente ou descendente)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalFound": {
                      "type": "integer",
                      "example": 20
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 4
                    },
                    "currentPage": {
                      "type": "integer",
                      "example": 1
                    },
                    "clients": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "example": "e7d66f9d-3c10-4e9e-b939-8a3a19159b6d"
                          },
                          "nomeCompleto": {
                            "type": "string",
                            "example": "alan Lima pereira"
                          },
                          "dataDeNascimento": {
                            "type": "string",
                            "format": "date",
                            "example": "1997-11-11"
                          },
                          "cpf": {
                            "type": "string",
                            "example": "321.654.987-22"
                          },
                          "email": {
                            "type": "string",
                            "format": "email",
                            "example": "marian111.lima95@yahoo.com"
                          },
                          "telefone": {
                            "type": "string",
                            "example": "11987654321"
                          },
                          "dataCadastro": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-10-13T17:39:38.000Z"
                          },
                          "dataExclusao": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true,
                            "example": "2024-12-15T12:30:00.000Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Requisição inválida. Verifique os parâmetros de entrada."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Falha na autenticação JWT (Token inválido ou ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Token inválido ou ausente"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor"
          }
        }
      }
    },
...
  },
...
}